requirements.txt
playwright==1.52.0
beautifulsoup4==4.13.4
chromadb==1.0.12
sentence-transformers==3.2.0
transformers==4.45.1
fastapi==0.115.9
uvicorn==0.34.3
streamlit==1.38.0
python-dotenv==1.1.0